+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/completion/AnyCompletionContributor.kt
@@ -18,52 +18,8 @@ import com.intellij.psi.PsiManager
 import com.intellij.psi.util.elementType
 import com.intellij.util.ProcessingContext
 import com.jetbrains.rd.util.string.printToString
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.*
 import org.apache.commons.lang3.StringUtils
-
-@Serializable
-data class PsiElementModelChild(
-    val model: PsiElementModel? = null,
-    var string: String? = null,
-    var array: Array<PsiElementModel?>? = null
-) {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
-
-        other as PsiElementModelChild
-
-        if (model != other.model) return false
-        if (string != other.string) return false
-        if (array != null) {
-            if (other.array == null) return false
-            if (!array.contentEquals(other.array)) return false
-        } else if (other.array != null) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = model?.hashCode() ?: 0
-        result = 31 * result + (string?.hashCode() ?: 0)
-        result = 31 * result + (array?.contentHashCode() ?: 0)
-        return result
-    }
-}
-
-@Serializable()
-data class PsiElementModel(
-    val id: String,
-    var elementType: String,
-    var options: MutableMap<String, PsiElementModelChild>,
-    var elementName: String?,
-    var elementFqn: String?,
-    var text: String?,
-    var parent: PsiElementModel?,
-    var prev: PsiElementModel?,
-    var next: PsiElementModel?
-)
 
 class AnyCompletionContributor() {
     class Completion : CompletionContributor() {
@@ -90,7 +46,8 @@ class AnyCompletionContributor() {
                             parameters.originalPosition,
                             parameters.originalFile,
                             RequestType.Completion,
-                            languageString
+                            languageString,
+                            parameters.offset
                         )
 
                         if (null === json) {
@@ -172,7 +129,8 @@ class AnyCompletionContributor() {
                     sourceElement,
                     sourceElement.containingFile,
                     RequestType.GoTo,
-                    languageString
+                    languageString,
+                    offset
                 )
 
             if (null === json) {