+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # intellij-psa Changelog
 
 ## [Unreleased]
+- Added `textRange` to the `PSA_CONTEXT` ENV variable. Shows PSI element position is the file
+- Added `PSA_OFFSET` ENV variable. Shows cursor position within current element in editor
 
 ## [0.0.5] - 2024-06-18
 - Performance optimizations