+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # intellij-psa Changelog
 
 ## [Unreleased]
+
+## [0.0.6] - 2024-06-22
 - Added `textRange` to the `PSA_CONTEXT` ENV variable. Shows PSI element position is the file
 - Added `PSA_OFFSET` ENV variable. Shows cursor position within current element in editor
 - Moved settings form to Kotlin UI DSL + improved it a lot
@@ -41,7 +43,8 @@
 ## [0.0.1] - 2024-06-09
 Initial Release
 
-[Unreleased]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/sam0delkin/intellij-psa/compare/v0.0.2...v0.0.3